+++ docs/asciidoc/modules/ROOT/pages/overview/apoc.coll/apoc.coll.avgDuration.adoc
@@ -5,7 +5,7 @@ This file is generated by DocsTest, so don't change it!
 = apoc.coll.avgDuration
 :description: This section contains reference documentation for the apoc.coll.avgDuration function.
 
-label:function[] label:apoc-core[]
+label:function[] label:apoc-full[]
 
 [.emphasis]
 apoc.coll.avgDuration([duration('P2DT3H'), duration('PT1H45S'), ...]) -  returns the average of a list of duration values
