+++ full/src/main/resources/extended.txt
@@ -159,6 +159,7 @@ apoc.uuid.install
 apoc.uuid.list
 apoc.uuid.remove
 apoc.uuid.removeAll
+apoc.coll.avgDuration
 apoc.data.email
 apoc.static.get
 apoc.static.getAll
